Update mastercode.m
button in the front, it stops car when pressed, need to work on it to make it turn left. ultrasonic sensor on right to detect openings, distance not finalized, also need to make it turn and then go forward a certain distance so it doesn't turn back the way it came from.